requirements.txt
anyjson==0.3.3
args==0.1.0
asgi-redis==1.4.2
asgiref==1.1.2
asn1crypto==0.22.0
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
Babel==2.5.1
backports-abc==0.5

backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
begins==0.9

billiard==3.3.0.23
cairocffi==0.8.0
CairoSVG==1.0.22

certifi==2017.7.27.1
cffi==1.10.0
channels==1.1.3
chardet==3.0.4

clint==0.5.1
constantly==15.1.0
cryptography==2.0.2
cssselect==1.0.1
cssselect2==0.2.0
cyordereddict==1.0.0
daphne==1.3.0
decorator==4.1.2
defusedxml==0.5.0
dict2xml==1.3
#dj-database-url==0.4.0
#Django==1.8.18
django-appconf==1.0.1
django-braces==1.11.0
django-celery==3.1.16
django-cors-headers==0.13
django-db-readonly==0.3.2
django-debug-toolbar==1.8
-e git+https://github.com/jnm/django-digest@709f932ab72078ce10cf5ff9917500cb6226dc1e#egg=django_digest
django-easy-pdf==0.1.1
django-excel==0.0.9
django-extensions==1.8.1
django-fcm==0.1.1
django-filter==0.7
#django-guardian==1.4.1
django-image-cropping==1.0.4
django-kombu==0.9.4
django-nose==1.4.2
django-oauth-toolkit==0.9.0
django-pure-pagination==0.3.0
#django-registration-redux==1.3
#django-reversion==1.8.4
django-snippetscream==0.0.7
django-statsd-mozilla==0.3.16
django-taggit==0.18.0
#django-templated-email==0.4.9
djangorestframework==3.3.2
djangorestframework-csv==1.4.1
djangorestframework-jsonp==1.0.2
djangorestframework-xml==1.3.0
docutils==0.13.1
dpath==1.2.post70
elaphe==0.5.6
enum34==1.1.6
ExifRead==2.1.2
Fabric==1.13.2
flower==0.9.2
#formpack==1.1
git+https://github.com/kobotoolbox/formpack@kobocat-requirement
funcsigs==1.0.2
futures==3.1.1
#gdata==2.0.18
gunicorn==19.7.1
html5lib==1.0b10
#httmock==1.2.2
#httplib2==0.10.3
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
jdcal==1.3
Jinja2==2.9.6
jsonfield==0.9.23
kombu==3.0.37
librabbitmq==1.5.2
linaro-django-pagination==2.0.3
lml==0.0.1
#lxml==3.4.0
Markdown==2.5
MarkupSafe==1.0
#mock==1.0.1
#modilabs-python-utils==0.1.5
msgpack-python==0.4.8
nose==1.3.7
numpy==1.13.1
oauthlib==1.0.1
olefile==0.44
openpyxl==2.0.5
pandas==0.20.3
paramiko==2.2.1
path.py==8.1.2
pathlib2==2.3.0
pdfkit==0.6.1
pexpect==4.2.1
pickleshare==0.7.4
#Pillow==6.2.0
#poster==0.8.1
preppy==2.4.1
prompt-toolkit==1.0.15
#psycopg2==2.5.4
ptyprocess==0.5.2
pyasn1==0.3.1
#pybamboo==0.5.8.1
pycparser==2.18
pyexcel==0.5.3
pyexcel-io==0.5.3
pyexcel-webio==0.1.2
pyexcel-xls==0.5.4
PyExcelerate==0.6.7
Pygments==2.2.0
#pymongo==2.7.2
PyNaCl==1.1.2
PyPDF2==1.26.0
Pyphen==0.9.4
pyquery==1.2.11
python-dateutil==2.6.1
python-digest==1.7
-e git+https://github.com/onaio/python-json2xlsclient.git@5a39387752d819cb6387f75569dbea9a5288aa6f#egg=python_json2xlsclient
#pytz==2017.2
#-e git+https://github.com/kobotoolbox/pyxform.git@d46d2edfd3bc1e1d25c06c3fd8aaef8324bc4f9f#egg=pyxform
#pyxform==0.11.5
recaptcha-client==1.0.6
redis==2.10.5
reportlab==3.4.0
requests==2.4.1
requests-oauthlib==0.8.0
#savReaderWriter==3.3.0
https://bitbucket.org/fomcl/savreaderwriter/downloads/savReaderWriter-3.3.0.zip#egg=savreaderwriter
scandir==1.5
selenium==3.8.0
shell-command==0.1
simplegeneric==0.8.1
simplejson==2.6.2
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.3
statistics==1.0.3.5
statsd==3.2.1
texttable==0.9.1
tinycss2==0.6.0
tornado==4.5.2
traitlets==4.3.2
transifex-client==0.12.4
twill==1.8.0
Twisted==17.5.0
txaio==2.8.1
typing==3.6.1
unicodecsv==0.9.4
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
WeasyPrint==0.40
webencodings==0.5.1
webstack-django-sorting==0.4.3
xhtml2pdf==0.2b1
xlwt==0.7.5
zope.interface==4.4.2
pyjwt==1.6.4
